+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Changelog
 
 Unreleased
 ==========
+
+1.0.4 (2022-04-05)
+==================
 * feat: Added a burger menu in the actions column of the ExtendedVersionAdminMixin.
 
 1.0.3 (2022-03-18)